Resolve bugs
